requirements.txt
Django==1.9
django-filter==0.7
djangorestframework
markdown==2.4.1
